requirements.txt
flask
pillow
jinja2
